+++ RELEASE_NOTES.md
@@ -1,8 +1,17 @@
 # Release Notes
 
+
+## Version 0.0.2
+
+    * Fineract AMS connector
+            * [SER-2000] - End payment hub transaction flow when an exception occurs
+            * [SER-2001] - Fineract AMS Connector throws error when amount is in decimal format
+
 ## Version 0.0.1
 
     * Fineract AMS connector
             * [SER-1568] - Project setup + Adding workers for Fineract verification and confirmation
             * [SER-1729] - Fix error that occurs when validation endpoint is called
             * [SER-1742] - Include loan ID in payload sent to fineract validation endpoint by the fineract AMS connector
+            * [SER-2000] - End payment hub transaction flow when an exception occurs
+            * [SER-2001] - Fineract AMS Connector throws error when amount is in decimal format